♻️ refactor : 휠 이벤트 함수 분리 #45
