+++ src/components/editor/editor-container.tsx
@@ -19,6 +19,7 @@ import TextCanvasElement from './elements/text/element-text-canvas';
 import UploadImageElement from './elements/uploads/element-upload-canvas';
 import TextEditContent from './elements/text/text-edit-content';
 import { SwitchCase } from '../common/switch-case';
+import { handleWheel } from '@/utils/editor/editor-scale-event.util';
 
 const EditorContainer = () => {
   const canvasElements = useEditorStore((state) => state.canvasElements);
@@ -38,7 +39,6 @@ const EditorContainer = () => {
   const backgroundColor = useEditorStore((state) => state.backgroundColor);
   const setSideBarStatus = sideBarStore((state) => state.setSideBarStatus);
   const zoom = sideBarStore((state) => state.zoom);
-  const setZoom = sideBarStore((state) => state.setZoom);
 
   //ref
   const transformerRef = useRef<Konva.Transformer | null>(null);
@@ -129,18 +129,6 @@ const EditorContainer = () => {
     if (!editingEl) return;
     updateElement(selectedElementId, { text: newText });
     setEditingElementId(null);
-  };
-
-  const handleWheel = (e: Konva.KonvaEventObject<WheelEvent>) => {
-    if (!e.evt.ctrlKey) return;
-    e.evt.preventDefault();
-
-    const scaleBy = 1.1;
-    const direction = e.evt.deltaY > 0 ? -1 : 1;
-    const newZoom = zoom * (direction > 0 ? scaleBy : 1 / scaleBy);
-
-    const clampedZoom = Math.min(Math.max(newZoom, 0.3), 3);
-    setZoom(clampedZoom);
   };
 
   return (
